+++ docs/misc/changelog.md
@@ -2,7 +2,7 @@
 
 # Changelog
 
-## Release 2.9.2a0 (2026-07-18)
+## Release 2.9.1a1 (2026-07-18)
 
 ### Breaking Changes: